add all features

+++ sakila-business-dao/src/main/java/isep/web/sakila/jpa/entities/Store.java
@@ -19,7 +19,7 @@ public class Store implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="store_id", unique=true, nullable=false)
-	private byte storeId;
+	private int storeId;
 
 	@Column(name="last_update", nullable=false)
 	private Timestamp lastUpdate;
@@ -49,7 +49,7 @@ public class Store implements Serializable {
 	public Store() {
 	}
 
-	public byte getStoreId() {
+	public int getStoreId() {
 		return this.storeId;
 	}
